% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMeanCI}
\alias{getMeanCI}
\title{Computes mean and confident interval for a vector of values}
\usage{
getMeanCI(var, nconf = 95)
}
\description{
Small function just to estimate mean, lower limit and upper limit for a given
set of values.
}
\keyword{internal}
