% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mse}
\alias{mse}
\title{Mean Squared Error}
\usage{
mse(actual, predicted)
}
\arguments{
\item{actual}{\code{numeric}. Vector of observed (true) values.}

\item{predicted}{\code{numeric}. Vector of predicted values. Must have the same
length as \code{actual}.}
}
\value{
A numeric value representing the mean squared error.
}
\description{
Computes the mean squared error (MSE) between observed (\code{actual}) and
predicted values. It is defined as the average of the squared differences
between both vectors.
}
\details{
The mean squared error is defined as:

\deqn{MSE = \frac{1}{n} \sum_{i=1}^{n} (actual_i - predicted_i)^2}

Lower values indicate better agreement between observed and predicted values.
}
\examples{
actual <- c(3, -0.5, 2, 7)
predicted <- c(2.5, 0.0, 2, 8)

mse(actual, predicted)

}
\keyword{internal}
