% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normScorePalette}
\alias{normScorePalette}
\title{Utility functions for internal use in normScore}
\usage{
normScorePalette(n)
}
\arguments{
\item{n}{Number of colors to retrieve}
}
\value{
A \code{vector} with n colors.
}
\description{
Serie of colors that represents the package, used for graphical representations.
}
\details{
This file contains helper functions used across the package,
including input validation, data formatting, and auxiliary computations.

These functions are not intended to be used directly by end users.

Just some colors for graphical representation
}
\keyword{internal}
