% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cv}
\alias{cv}
\title{Coefficient of Variation}
\usage{
cv(x, proportion = TRUE, na.rm = TRUE)
}
\arguments{
\item{x}{\code{numeric}. Numeric vector for which the coefficient of variation
will be calculated.}

\item{proportion}{\code{logical}. If \code{TRUE}, the result is returned as a proportion.
If \code{FALSE}, it is multiplied by 100 and returned as a percentage.
Default is \code{TRUE}.}

\item{na.rm}{\code{logical}. Should missing values (\code{NA}) be removed before
computing the mean and standard deviation? Default is \code{TRUE}.}
}
\value{
A numeric value representing the coefficient of variation of \code{x}.
}
\description{
Computes the coefficient of variation (CV) of a numeric vector as the ratio
between the standard deviation and the mean. The result can be returned as
a proportion or as a percentage.
}
\details{
The coefficient of variation is defined as:

\deqn{CV = \frac{sd(x)}{mean(x)}}

If \code{proportion = FALSE}, the result is multiplied by 100.

Note that if the mean of \code{x} is 0 or very close to 0, the result may be
\code{Inf}, \code{NaN}, or unstable.
}
\keyword{internal}
