% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isSameNormalization}
\alias{isSameNormalization}
\title{Compare Two Normalized Datasets}
\usage{
isSameNormalization(x, y, tolerance = 1e-08)
}
\arguments{
\item{x}{First normalized data matrix or data frame.}

\item{y}{Second normalized data matrix or data frame.}

\item{tolerance}{\code{numeric}. Numerical tolerance used for comparison.
Default is \code{1e-8}.}
}
\value{
A \code{logical} value indicating whether \code{x} and \code{y} are considered equivalent.
}
\description{
Checks whether two normalized datasets are equivalent, allowing for small
numerical differences.
}
\keyword{internal}
