% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mape}
\alias{mape}
\title{Mean Absolute Percentage Error}
\usage{
mape(actual, predicted, proportion = FALSE)
}
\arguments{
\item{actual}{\code{numeric}. Vector of observed (true) values.}

\item{predicted}{\code{numeric}. Vector of predicted values. Must have the same
length as \code{actual}.}

\item{proportion}{\code{logical}. If \code{TRUE}, the result is returned as a proportion.
If \code{FALSE}, it is multiplied by 100 and returned as a percentage.
Default is \code{FALSE}.}
}
\value{
A numeric value representing the mean absolute percentage error.
}
\description{
Computes the mean absolute percentage error (MAPE) between observed
(\code{actual}) and predicted values. The result can be returned as a proportion
or as a percentage.
}
\details{
The mean absolute percentage error is defined as:

\deqn{MAPE = \frac{1}{n} \sum_{i=1}^{n} \left| \frac{actual_i - predicted_i}{actual_i} \right|}

If \code{proportion = FALSE}, the result is multiplied by 100.

Note that this metric is undefined when any value in \code{actual} is equal to 0.
In such cases, the function may return \code{Inf} or \code{NaN}.
}
\examples{
actual <- c(100, 200, 300, 400)
predicted <- c(110, 190, 320, 390)

# As percentage
mape(actual, predicted)

# As proportion
mape(actual, predicted, proportion = TRUE)

}
\keyword{internal}
